get that working on windows too
