+++ Code/GraphMol/Wrap/Chirality.cpp
@@ -17,10 +17,6 @@
 
 namespace python = boost::python;
 
-void setAllowNontetrahedral(bool val) {
-  RDKit::Chirality::allowNontetrahedralChirality = val;
-}
-
 namespace RDKit {
 struct chirality_wrapper {
   static void wrap() {
@@ -62,9 +58,6 @@ struct chirality_wrapper {
         .def_readonly("controllingAtoms",
                       &Chirality::StereoInfo::controllingAtoms,
                       "indices of the atoms controlling the stereo");
-    python::def(
-        "SetAllowNontetrahedralChirality", setAllowNontetrahedral,
-        "toggles recognition of non-tetrahedral chirality from 3D structures");
   };
 };
 }  // namespace RDKit
