+++ Code/GraphMol/Wrap/MolOps.cpp
@@ -33,12 +33,18 @@
 #include <RDBoost/PySequenceHolder.h>
 #include <RDBoost/Wrap.h>
 #include <RDBoost/python_streambuf.h>
+#include <GraphMol/Chirality.h>
 
 #include <sstream>
 namespace python = boost::python;
 using boost_adaptbx::python::streambuf;
 
 namespace RDKit {
+
+void setAllowNontetrahedral(bool val) {
+  RDKit::Chirality::allowNontetrahedralChirality = val;
+}
+
 python::tuple fragmentOnSomeBondsHelper(const ROMol &mol,
                                         python::object pyBondIndices,
                                         unsigned int nToBreak, bool addDummies,
@@ -913,14 +919,13 @@ void setDoubleBondNeighborDirectionsHelper(ROMol &mol, python::object confObj) {
 void setAtomSymbols(MolzipParams &p, python::object symbols) {
   p.atomSymbols.clear();
   if (symbols) {
-    unsigned int nVs =
-        python::extract<unsigned int>(symbols.attr("__len__")());
+    unsigned int nVs = python::extract<unsigned int>(symbols.attr("__len__")());
     for (unsigned int i = 0; i < nVs; ++i) {
       p.atomSymbols.push_back(python::extract<std::string>(symbols[i]));
     }
   }
 }
-  
+
 ROMol *molzip_new(const ROMol &a, const ROMol &b, const MolzipParams &p) {
   return molzip(a, b, p).release();
 }
@@ -2505,12 +2510,13 @@ EXAMPLES:\n\n\
                                  python::init<>())
         .def_readwrite("label", &MolzipParams::label,
                        "Set the atom labelling system to zip together")
-        .def_readwrite("enforceValenceRules", &MolzipParams::enforceValenceRules,
-		       "If true (default) enforce valences after zipping\n\
+        .def_readwrite("enforceValenceRules",
+                       &MolzipParams::enforceValenceRules,
+                       "If true (default) enforce valences after zipping\n\
 Setting this to false allows assembling chemically incorrect fragments.")
         .def("setAtomSymbols", &RDKit::setAtomSymbols,
-             "Set the atom symbols used to zip mols together when using AtomType labeling")
-      ;
+             "Set the atom symbols used to zip mols together when using "
+             "AtomType labeling");
 
     docString =
         "molzip: zip two molecules together preserving bond and atom stereochemistry.\n\
@@ -2723,6 +2729,9 @@ A note on the flags controlling which atoms/bonds are modified:
     python::def("ConvertGenericQueriesToSubstanceGroups",
                 GenericGroups::convertGenericQueriesToSubstanceGroups,
                 python::arg("mol"), "documentation");
+    python::def(
+        "SetAllowNontetrahedralChirality", setAllowNontetrahedral,
+        "toggles recognition of non-tetrahedral chirality from 3D structures");
   }
 };
 }  // namespace RDKit
